+++ packages/test-base/src/jvmMain/kotlin/io/realm/kotlin/test/platform/PlatformUtils.kt
@@ -23,6 +23,7 @@ import java.nio.file.attribute.PosixFilePermission
 import java.util.stream.Collectors
 import kotlin.io.path.absolutePathString
 import kotlin.time.Duration
+import kotlin.time.Duration.Companion.seconds
 
 actual object PlatformUtils {
     actual fun createTempDir(prefix: String, readOnly: Boolean): String {
@@ -45,16 +46,27 @@ actual object PlatformUtils {
         val pathsToDelete: List<Path> =
             Files.walk(rootPath).sorted(Comparator.reverseOrder()).collect(Collectors.toList())
         for (p in pathsToDelete) {
-            try {
-                Files.deleteIfExists(p)
-            } catch (e: java.nio.file.FileSystemException) {
-                // Sometimes (on Windows) we need the give a GC a chance to run and close all native pointers
-                // before we can delete the Realm, otherwise delete will fail with " The process cannot access the
-                // file because it is being used by another process".
-                //
-                // We try to trigger the GC once then retry the delete.
-                triggerGC()
-                Files.deleteIfExists(p)
+            // Sometimes (on Windows) we need the give a GC a chance to run and close all native pointers
+            // before we can delete the Realm, otherwise delete will fail with " The process cannot access the
+            // file because it is being used by another process".
+            //
+            // We try to trigger the GC once then retry the delete.
+            var counter = 5
+            var deleted = false
+            var error: java.nio.file.FileSystemException? = null
+            while (!deleted && counter > 0) {
+                try {
+                    Files.deleteIfExists(p)
+                    deleted = true
+                } catch (e: java.nio.file.FileSystemException) {
+                    error = e
+                    triggerGC()
+                    sleep(1.seconds)
+                    counter -= 1
+                }
+            }
+            if (!deleted) {
+                throw error!!
             }
         }
     }